Broken commit: Extract folding API code from PoseEditMode
This commit didn't modify the copied code, only the surrounding code, which is well visible in the diff.
